requirements.txt
yfinance==0.2.32
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.8.5
plotly==5.14.0
streamlit==1.25.0
ta==0.10.2
beautifulsoup4==4.12.2
lxml==4.9.2
alpaca-trade-api==3.0.0
